backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
python-multipart>=0.0.5
pydantic>=1.9.0
pymupdf>=1.19.0
langchain>=0.0.267
langchain-community>=0.0.10
langchain-core>=0.1.1
langchain-google-genai>=0.0.5
python-dotenv>=0.19.0
google-generativeai>=0.3.1
